# contratos/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "auto_cusco API",
    "description": "API REST del backend de auto_cusco: ingesta de sabanas, seleccion y segmentacion de cartera, generacion de cargas (digitales y VoIP) y reportes de cobranza.",
    "version": "0.1.0"
  },
  "paths": {
...
    "/cargas": {
      "post": {
        "tags": [
          "cargas"
        ],
        "summary": "Subir Sabana",
        "description": "Registra una version nueva y lanza su procesamiento en segundo plano (V-1).",
        "operationId": "subir_sabana_cargas_post",
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_subir_sabana_cargas_post"
              }
            }
          }
        },
        "responses": {
          "202": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CargaCreadaRespuesta"
                }
              }
            }
          },
          "400": {
            "description": "La peticion no se puede atender tal como viene; el motivo esta en detail",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DetalleError"
                }
              }
            }
          },
          "413": {
            "description": "El archivo supera el tamano maximo",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DetalleError"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "cargas"
        ],
        "summary": "Listar Versiones",
        "operationId": "listar_versiones_cargas_get",
        "parameters": [
          {
            "name": "fecha_corte",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "date"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Fecha Corte"
            }
          },
          {
            "name": "limite",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 200,
              "minimum": 1,
              "default": 50,
              "title": "Limite"
            }
          },
          {
            "name": "desplazamiento",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0,
              "title": "Desplazamiento"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/VersionRespuesta"
                  },
                  "title": "Response Listar Versiones Cargas Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/cargas/{carga_id}": {
      "get": {
        "tags": [
          "cargas"
        ],
        "summary": "Obtener Version",
        "operationId": "obtener_version_cargas__carga_id__get",
        "parameters": [
          {
            "name": "carga_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Carga Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VersionRespuesta"
                }
              }
            }
          },
          "404": {
            "description": "No existe lo pedido",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DetalleError"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "cargas"
        ],
        "summary": "Eliminar Version",
        "operationId": "eliminar_version_cargas__carga_id__delete",
        "parameters": [
          {
            "name": "carga_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Carga Id"
            }
          },
          {
            "name": "dejar_fecha_sin_vigente",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "description": "Confirma que la fecha quede sin version vigente (V-8)",
              "default": false,
              "title": "Dejar Fecha Sin Vigente"
            },
            "description": "Confirma que la fecha quede sin version vigente (V-8)"
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "404": {
            "description": "No existe lo pedido",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DetalleError"
                }
              }
            }
          },
          "409": {
            "description": "La operacion choca con el estado actual",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DetalleError"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/cargas/{carga_id}/incidencias": {
      "get": {
        "tags": [
          "cargas"
        ],
        "summary": "Listar Incidencias",
        "operationId": "listar_incidencias_cargas__carga_id__incidencias_get",
        "parameters": [
          {
            "name": "carga_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Carga Id"
            }
          },
          {
            "name": "severidad",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/Severidad"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Severidad"
            }
          },
          {
            "name": "limite",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 500,
              "minimum": 1,
              "default": 100,
              "title": "Limite"
            }
          },
          {
            "name": "desplazamiento",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0,
              "title": "Desplazamiento"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IncidenciasRespuesta"
                }
              }
            }
          },
          "404": {
            "description": "No existe lo pedido",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DetalleError"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/cargas/{carga_id}/vigente": {
      "post": {
        "tags": [
          "cargas"
        ],
        "summary": "Asignar Vigente",
        "description": "Regla V-7: elige que version se considera para su fecha.",
        "operationId": "asignar_vigente_cargas__carga_id__vigente_post",
        "parameters": [
          {
            "name": "carga_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Carga Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VersionRespuesta"
                }
              }
            }
          },
          "404": {
            "description": "No existe lo pedido",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DetalleError"
                }
              }
            }
          },
          "409": {
            "description": "La operacion choca con el estado actual",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DetalleError"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Body_subir_sabana_cargas_post": {
        "properties": {
          "fecha_corte": {
            "type": "string",
            "format": "date",
            "title": "Fecha Corte",
            "description": "Fecha del corte que representa la sabana"
          },
          "archivo": {
            "type": "string",
            "contentMediaType": "application/octet-stream",
            "title": "Archivo",
            "description": "Archivo de la sabana (.xlsb, .xlsx)"
          },
          "hoja": {
            "type": "string",
            "title": "Hoja",
            "default": "VENCIDA"
          }
        },
        "type": "object",
        "required": [
          "fecha_corte",
          "archivo"
        ],
        "title": "Body_subir_sabana_cargas_post"
      },
...
      "CargaCreadaRespuesta": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "fecha_corte": {
            "type": "string",
            "format": "date",
            "title": "Fecha Corte"
          },
          "version": {
            "type": "integer",
            "title": "Version"
          },
          "estado": {
            "$ref": "#/components/schemas/EstadoCarga"
          },
          "versiones_identicas": {
            "items": {
              "type": "integer"
            },
            "type": "array",
            "title": "Versiones Identicas"
          },
          "aviso": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Aviso"
          }
        },
        "type": "object",
        "required": [
          "id",
          "fecha_corte",
          "version",
          "estado",
          "versiones_identicas",
          "aviso"
        ],
        "title": "CargaCreadaRespuesta"
      },
...
      "DetalleError": {
        "properties": {
          "detail": {
            "type": "string",
            "title": "Detail"
          }
        },
        "type": "object",
        "required": [
          "detail"
        ],
        "title": "DetalleError",
        "description": "Motivo legible por el que la API rechazo la peticion."
      },
...
      "EstadoCarga": {
        "type": "string",
        "enum": [
          "en_cola",
          "procesando",
          "terminada",
          "fallida"
        ],
        "title": "EstadoCarga",
        "description": "Estado de procesamiento de una version (la ingesta corre en segundo plano)."
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "IncidenciaRespuesta": {
        "properties": {
          "fila": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Fila"
          },
          "columna": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Columna"
          },
          "codigo": {
            "type": "string",
            "title": "Codigo"
          },
          "severidad": {
            "$ref": "#/components/schemas/Severidad"
          },
          "detalle": {
            "type": "string",
            "title": "Detalle"
          },
          "valor_original": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Valor Original"
          }
        },
        "type": "object",
        "required": [
          "fila",
          "columna",
          "codigo",
          "severidad",
          "detalle",
          "valor_original"
        ],
        "title": "IncidenciaRespuesta"
      },
      "IncidenciasRespuesta": {
        "properties": {
          "total": {
            "type": "integer",
            "title": "Total"
          },
          "incidencias": {
            "items": {
              "$ref": "#/components/schemas/IncidenciaRespuesta"
            },
            "type": "array",
            "title": "Incidencias"
          }
        },
        "type": "object",
        "required": [
          "total",
          "incidencias"
        ],
        "title": "IncidenciasRespuesta"
      },
...
      "Severidad": {
        "type": "string",
        "enum": [
          "error",
          "advertencia",
          "info"
        ],
        "title": "Severidad",
        "description": "Gravedad de una incidencia detectada al normalizar."
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
      "VersionRespuesta": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "fecha_corte": {
            "type": "string",
            "format": "date",
            "title": "Fecha Corte"
          },
          "version": {
            "type": "integer",
            "title": "Version"
          },
          "estado": {
            "$ref": "#/components/schemas/EstadoCarga"
          },
          "vigente": {
            "type": "boolean",
            "title": "Vigente"
          },
          "nombre_archivo": {
            "type": "string",
            "title": "Nombre Archivo"
          },
          "tamano_bytes": {
            "type": "integer",
            "title": "Tamano Bytes"
          },
          "hoja": {
            "type": "string",
            "title": "Hoja"
          },
          "creado_en": {
            "type": "string",
            "format": "date-time",
            "title": "Creado En"
          },
          "procesado_en": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Procesado En"
          },
          "huella_formato": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Huella Formato"
          },
          "filas_total": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Filas Total"
          },
          "filas_ingestadas": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Filas Ingestadas"
          },
          "incidencias_error": {
            "type": "integer",
            "title": "Incidencias Error",
            "default": 0
          },
          "incidencias_advertencia": {
            "type": "integer",
            "title": "Incidencias Advertencia",
            "default": 0
          },
          "incidencias_info": {
            "type": "integer",
            "title": "Incidencias Info",
            "default": 0
          },
          "motivo_fallo": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Motivo Fallo"
          }
        },
        "type": "object",
        "required": [
          "id",
          "fecha_corte",
          "version",
          "estado",
          "vigente",
          "nombre_archivo",
          "tamano_bytes",
          "hoja",
          "creado_en",
          "procesado_en",
          "huella_formato",
          "filas_total",
          "filas_ingestadas",
          "incidencias_error",
          "incidencias_advertencia",
          "incidencias_info",
          "motivo_fallo"
        ],
        "title": "VersionRespuesta"
      }
    }
  }
}